requirements.txt
torch
numpy
pandas
scikit-learn
pyyaml
tqdm
onnx
onnxruntime
matplotlib
